move atomic list
