+++ modules/bridge/src/main/kotlin/ru/astrainteractive/discordbot/module/bridge/WebSocketClient.kt
@@ -24,7 +24,7 @@ import okio.ByteString
 import ru.astrainteractive.astralibs.async.CoroutineFeature
 import ru.astrainteractive.astralibs.logging.JUtiltLogger
 import ru.astrainteractive.astralibs.logging.Logger
-import ru.astrainteractive.discordbot.module.bridge.model.AtomicList
+import ru.astrainteractive.discordbot.module.bridge.util.AtomicList
 import ru.astrainteractive.discordbot.module.bridge.model.SocketMessage
 import ru.astrainteractive.discordbot.module.bridge.model.SocketRoute
 import ru.astrainteractive.discordbot.module.bridge.serializer.SocketMessageFactory
